refactor: remove default export in order to export types
refs: CDS-26
